Refactor #68 : MapFrame 컴포넌트 분리
- Home 에서 사용되던 Map을 MapFrame 컴포넌트로 분리후 재사용
- 헤더 추가
