+++ client/src/components/MapFrame/index.ts
@@ -1,0 +1,1 @@
+export { default } from './MapFrame';
